gui: reworking heat maps into interfaces other tools can extend
