+++ src/gui/CMakeLists.txt
@@ -46,6 +46,8 @@ if (Qt5_FOUND AND BUILD_GUI)
     src/drcWidget.cpp
     src/ruler.cpp
     src/heatMap.cpp
+    src/heatMapSetup.cpp
+    src/heatMapPlacementDensity.cpp
     resources/resource.qrc
   )
 
@@ -57,7 +59,6 @@ if (Qt5_FOUND AND BUILD_GUI)
       Qt5::Widgets
       utl
       Boost::boost
-      psm
   )
 
 messages(
@@ -66,7 +67,10 @@ messages(
 
 else()
   message(STATUS "GUI is not enabled")
-  add_library(gui src/stub.cpp)
+  add_library(gui 
+    src/stub.cpp 
+    src/stub_heatMap.cpp
+  )
   target_link_libraries(gui
     PUBLIC
       odb
