+++ src/gui/src/gui.cpp
@@ -54,6 +54,7 @@
 
 #include "drcWidget.h"
 #include "ruler.h"
+#include "heatMapPlacementDensity.h"
 
 extern int cmd_argc;
 extern char **cmd_argv;
@@ -76,6 +77,9 @@ static odb::dbBlock* getBlock(odb::dbDatabase* db)
 
 // This provides the link for Gui::redraw to the widget
 static gui::MainWindow* main_window = nullptr;
+
+// Placement density heatmap
+static std::unique_ptr<PlacementDensityDataSource> placement_density_heat_map = nullptr;
 
 // Used by toString to convert dbu to microns (and back), will be set in main_window
 DBUToString Descriptor::Property::convert_dbu;
@@ -121,7 +125,9 @@ bool Gui::enabled()
 
 void Gui::registerRenderer(Renderer* renderer)
 {
-  main_window->getControls()->registerRenderer(renderer);
+  if (Gui::enabled()) {
+    main_window->getControls()->registerRenderer(renderer);
+  }
 
   renderers_.insert(renderer);
   redraw();
@@ -542,9 +548,104 @@ void Gui::showWidget(const std::string& name, bool show)
   }
 }
 
+void Gui::registerHeatMap(HeatMapDataSource* heatmap)
+{
+  heat_maps_.insert(heatmap);
+  registerRenderer(heatmap->getRenderer());
+  if (Gui::enabled()) {
+    main_window->registerHeatMap(heatmap);
+  }
+}
+
+void Gui::unregisterHeatMap(HeatMapDataSource* heatmap)
+{
+  if (heat_maps_.count(heatmap) == 0) {
+      return;
+  }
+
+  unregisterRenderer(heatmap->getRenderer());
+  if (Gui::enabled()) {
+    main_window->unregisterHeatMap(heatmap);
+  }
+  heat_maps_.erase(heatmap);
+}
+
 void Gui::setHeatMapSetting(const std::string& name, const std::string& option, const Renderer::Setting& value)
 {
-  main_window->setHeatMapSetting(name, option, value);
+  HeatMapDataSource* source = nullptr;
+
+  for (auto* heat_map : heat_maps_) {
+    if (heat_map->getShortName() == name) {
+      source = heat_map;
+      break;
+    }
+  }
+
+  if (source == nullptr) {
+    QStringList options;
+    for (auto* heat_map : heat_maps_) {
+      options.append(QString::fromStdString(heat_map->getShortName()));
+    }
+    logger_->error(utl::GUI, 28, "{} is not a known map. Valid options are: {}", name, options.join(", ").toStdString());
+  }
+
+  const std::string rebuild_map_option = "rebuild";
+  if (option == rebuild_map_option) {
+    source->destroyMap();
+  } else {
+    auto settings = source->getSettings();
+
+    if (settings.count(option) == 0) {
+      QStringList options;
+      options.append(QString::fromStdString(rebuild_map_option));
+      for (const auto& [key, kv] : settings) {
+        options.append(QString::fromStdString(key));
+      }
+      logger_->error(utl::GUI, 29, "{} is not a valid option. Valid options are: {}", option, options.join(", ").toStdString());
+    }
+
+    auto current_value = settings[option];
+    if (std::holds_alternative<bool>(current_value)) {
+      // is bool
+      if (auto* s = std::get_if<bool>(&value)) {
+        settings[option] = *s;
+      } if (auto* s = std::get_if<int>(&value)) {
+        settings[option] = *s != 0;
+      } if (auto* s = std::get_if<double>(&value)) {
+        settings[option] = *s != 0.0;
+      } else {
+        logger_->error(utl::GUI, 60, "{} must be a boolean", option);
+      }
+    } else if (std::holds_alternative<int>(current_value)) {
+      // is int
+      if (auto* s = std::get_if<int>(&value)) {
+        settings[option] = *s;
+      } else if (auto* s = std::get_if<double>(&value)) {
+        settings[option] = static_cast<int>(*s);
+      } else {
+        logger_->error(utl::GUI, 61, "{} must be an integer or double", option);
+      }
+    } else if (std::holds_alternative<double>(current_value)) {
+      // is double
+      if (auto* s = std::get_if<int>(&value)) {
+        settings[option] = static_cast<double>(*s);
+      } else if (auto* s = std::get_if<double>(&value)) {
+        settings[option] = *s;
+      } else {
+        logger_->error(utl::GUI, 62, "{} must be an integer or double", option);
+      }
+    } else {
+      // is string
+      if (auto* s = std::get_if<std::string>(&value)) {
+        settings[option] = *s;
+      } else {
+        logger_->error(utl::GUI, 63, "{} must be a string", option);
+      }
+    }
+    source->setSettings(settings);
+  }
+
+  source->getRenderer()->redraw();
 }
 
 Renderer::~Renderer()
@@ -809,7 +910,7 @@ int startGui(int& argc, char* argv[], Tcl_Interp* interp, const std::string& scr
   main_window->getScriptWidget()->setupTcl(interp, init_openroad);
 
   // openroad is guaranteed to be initialized here
-  main_window->init(open_road->getSta(), open_road->getPDNSim());
+  main_window->init(open_road->getSta());
 
   // Exit the app if someone chooses exit from the menu in the window
   QObject::connect(main_window, SIGNAL(exit()), &app, SLOT(quit()));
@@ -990,6 +1091,10 @@ void initGui(OpenRoad* openroad)
   auto* gui = gui::Gui::get();
   gui->setDatabase(openroad->getDb());
   gui->setLogger(openroad->getLogger());
+
+  // placement density heatmap
+  gui::placement_density_heat_map = std::make_unique<gui::PlacementDensityDataSource>(openroad->getLogger());
+  gui::placement_density_heat_map->registerHeatMap();
 }
 
 }  // namespace ord
