+++ src/gui/src/heatMap.cpp
@@ -30,8 +30,9 @@
 // ARISING IN ANY WAY OUT OF THE USE OF THIS SOFTWARE, EVEN IF ADVISED OF THE
 // POSSIBILITY OF SUCH DAMAGE.
 
-#include "heatMap.h"
-#include "db_sta/dbNetwork.hh"
+#include "utl/Logger.h"
+#include "gui/heatMap.h"
+#include "heatMapSetup.h"
 
 #include <QComboBox>
 #include <QHBoxLayout>
@@ -43,294 +44,8 @@
 
 namespace gui {
 
-HeatMapSetup::HeatMapSetup(HeatMapDataSource& source,
-                           const QString& title,
-                           bool use_dbu,
-                           int dbu,
-                           QWidget* parent) :
-    QDialog(parent),
-    source_(source),
-    use_dbu_(use_dbu),
-    dbu_(dbu),
-    log_scale_(new QCheckBox(this)),
-    reverse_log_scale_(new QCheckBox(this)),
-    show_numbers_(new QCheckBox(this)),
-    show_legend_(new QCheckBox(this)),
-    grid_x_size_(nullptr),
-    grid_y_size_(nullptr),
-    grid_x_size_dbu_(nullptr),
-    grid_y_size_dbu_(nullptr),
-    min_range_selector_(new QDoubleSpinBox(this)),
-    show_mins_(new QCheckBox(this)),
-    max_range_selector_(new QDoubleSpinBox(this)),
-    show_maxs_(new QCheckBox(this)),
-    alpha_selector_(new QSpinBox(this)),
-    rebuild_(new QPushButton("Rebuild data", this)),
-    close_(new QPushButton("Close", this))
-{
-  setWindowTitle(title);
-
-  QVBoxLayout* overall_layout = new QVBoxLayout;
-  QFormLayout* form = new QFormLayout;
-
-  source_.makeAdditionalSetupOptions(this, form, [this]() { source_.redraw(); });
-
-  form->addRow(tr("Log scale"), log_scale_);
-  form->addRow(tr("Reverse Log scale"), reverse_log_scale_);
-
-  form->addRow(tr("Show numbers"), show_numbers_);
-
-  form->addRow(tr("Show legend"), show_legend_);
-
-  QHBoxLayout* grid_layout = new QHBoxLayout;
-  if (!use_dbu_) {
-    grid_x_size_ = new QDoubleSpinBox(this);
-    grid_y_size_ = new QDoubleSpinBox(this);
-
-    const QString grid_suffix(" \u03BCm"); // micro meters
-    grid_x_size_->setRange(source_.getGridSizeMinimumValue(), source_.getGridSizeMaximumValue());
-    grid_x_size_->setSuffix(grid_suffix);
-    grid_y_size_->setRange(source_.getGridSizeMinimumValue(), source_.getGridSizeMaximumValue());
-    grid_y_size_->setSuffix(grid_suffix);
-    grid_layout->addWidget(new QLabel("X", this));
-    grid_layout->addWidget(grid_x_size_);
-    grid_layout->addWidget(new QLabel("Y", this));
-    grid_layout->addWidget(grid_y_size_);
-    if (!source_.canAdjustGrid()) {
-      grid_x_size_->setEnabled(false);
-      grid_y_size_->setEnabled(false);
-    }
-  } else {
-    grid_x_size_dbu_ = new QSpinBox(this);
-    grid_y_size_dbu_ = new QSpinBox(this);
-
-    grid_x_size_dbu_->setRange(source_.getGridSizeMinimumValue() * dbu_, source_.getGridSizeMaximumValue() * dbu_);
-    grid_y_size_dbu_->setRange(source_.getGridSizeMinimumValue() * dbu_, source_.getGridSizeMaximumValue() * dbu_);
-    grid_layout->addWidget(new QLabel("X", this));
-    grid_layout->addWidget(grid_x_size_dbu_);
-    grid_layout->addWidget(new QLabel("Y", this));
-    grid_layout->addWidget(grid_y_size_dbu_);
-    if (!source_.canAdjustGrid()) {
-      grid_x_size_dbu_->setEnabled(false);
-      grid_y_size_dbu_->setEnabled(false);
-    }
-  }
-  form->addRow(tr("Grid"), grid_layout);
-
-  min_range_selector_->setDecimals(3);
-  QHBoxLayout* min_grid = new QHBoxLayout;
-  min_grid->addWidget(min_range_selector_);
-  min_grid->addWidget(new QLabel("Show values below", this));
-  min_grid->addWidget(show_mins_);
-  form->addRow(tr("Minimum"), min_grid);
-
-  max_range_selector_->setDecimals(3);
-  QHBoxLayout* max_grid = new QHBoxLayout;
-  max_grid->addWidget(max_range_selector_);
-  max_grid->addWidget(new QLabel("Show values above", this));
-  max_grid->addWidget(show_maxs_);
-  form->addRow(tr("Maximum"), max_grid);
-
-  alpha_selector_->setRange(source_.getColorAlphaMinimum(), source_.getColorAlphaMaximum());
-  form->addRow(tr("Color alpha"), alpha_selector_);
-
-  overall_layout->addLayout(form);
-
-  QHBoxLayout* buttons = new QHBoxLayout;
-  buttons->addWidget(rebuild_);
-  buttons->addWidget(close_);
-
-  overall_layout->addLayout(buttons);
-
-  setLayout(overall_layout);
-
-  updateWidgets();
-
-  connect(log_scale_,
-          SIGNAL(stateChanged(int)),
-          this,
-          SLOT(updateScale(int)));
-
-  connect(reverse_log_scale_,
-          SIGNAL(stateChanged(int)),
-          this,
-          SLOT(updateReverseScale(int)));
-
-  if (!use_dbu_) {
-    connect(grid_x_size_,
-            SIGNAL(valueChanged(double)),
-            this,
-            SLOT(updateGridSize()));
-    connect(grid_y_size_,
-            SIGNAL(valueChanged(double)),
-            this,
-            SLOT(updateGridSize()));
-  } else {
-    connect(grid_x_size_dbu_,
-            SIGNAL(valueChanged(int)),
-            this,
-            SLOT(updateGridSize()));
-    connect(grid_y_size_dbu_,
-            SIGNAL(valueChanged(int)),
-            this,
-            SLOT(updateGridSize()));
-  }
-
-  connect(show_numbers_,
-          SIGNAL(stateChanged(int)),
-          this,
-          SLOT(updateShowNumbers(int)));
-
-  connect(show_legend_,
-          SIGNAL(stateChanged(int)),
-          this,
-          SLOT(updateShowLegend(int)));
-
-  connect(min_range_selector_,
-          SIGNAL(valueChanged(double)),
-          this,
-          SLOT(updateRange()));
-  connect(max_range_selector_,
-          SIGNAL(valueChanged(double)),
-          this,
-          SLOT(updateRange()));
-  connect(show_mins_,
-          SIGNAL(stateChanged(int)),
-          this,
-          SLOT(updateShowMinRange(int)));
-  connect(show_maxs_,
-          SIGNAL(stateChanged(int)),
-          this,
-          SLOT(updateShowMaxRange(int)));
-
-  connect(alpha_selector_,
-          SIGNAL(valueChanged(int)),
-          this,
-          SLOT(updateAlpha(int)));
-
-  connect(this,
-          SIGNAL(changed()),
-          this,
-          SLOT(updateWidgets()));
-
-  connect(rebuild_,
-          SIGNAL(pressed()),
-          this,
-          SLOT(destroyMap()));
-
-  connect(close_,
-          SIGNAL(pressed()),
-          this,
-          SLOT(accept()));
-}
-
-void HeatMapSetup::updateWidgets()
-{
-  show_mins_->setCheckState(source_.getDrawBelowRangeMin() ? Qt::Checked : Qt::Unchecked);
-
-  min_range_selector_->blockSignals(true);
-  min_range_selector_->setRange(source_.convertPercentToValue(source_.getDisplayRangeMinimumValue()),
-                                source_.convertPercentToValue(source_.getDisplayRangeMaximumValue()));
-  min_range_selector_->setValue(source_.convertPercentToValue(source_.getDisplayRangeMin()));
-  min_range_selector_->blockSignals(false);
-  min_range_selector_->setSuffix(" " + QString::fromStdString(source_.getValueUnits()));
-  min_range_selector_->setSingleStep(source_.getDisplayRangeIncrement());
-
-  show_maxs_->setCheckState(source_.getDrawAboveRangeMax() ? Qt::Checked : Qt::Unchecked);
-
-  max_range_selector_->blockSignals(true);
-  max_range_selector_->setRange(source_.convertPercentToValue(source_.getDisplayRangeMinimumValue()),
-                                source_.convertPercentToValue(source_.getDisplayRangeMaximumValue()));
-  max_range_selector_->setValue(source_.convertPercentToValue(source_.getDisplayRangeMax()));
-  max_range_selector_->blockSignals(false);
-  max_range_selector_->setSuffix(" " + QString::fromStdString(source_.getValueUnits()));
-  max_range_selector_->setSingleStep(source_.getDisplayRangeIncrement());
-
-  if (!use_dbu_) {
-    grid_x_size_->setValue(source_.getGridXSize());
-    grid_y_size_->setValue(source_.getGridYSize());
-  } else {
-    grid_x_size_dbu_->setValue(source_.getGridXSize() * dbu_);
-    grid_y_size_dbu_->setValue(source_.getGridYSize() * dbu_);
-  }
-
-  alpha_selector_->setValue(source_.getColorAlpha());
-
-  log_scale_->setCheckState(source_.getLogScale() ? Qt::Checked : Qt::Unchecked);
-  reverse_log_scale_->setCheckState(source_.getReverseLogScale() ? Qt::Checked : Qt::Unchecked);
-  reverse_log_scale_->setEnabled(source_.getLogScale());
-  show_numbers_->setCheckState(source_.getShowNumbers() ? Qt::Checked : Qt::Unchecked);
-  show_legend_->setCheckState(source_.getShowLegend() ? Qt::Checked : Qt::Unchecked);
-}
-
-void HeatMapSetup::destroyMap()
-{
-  source_.destroyMap();
-}
-
-void HeatMapSetup::updateScale(int option)
-{
-  source_.setLogScale(option == Qt::Checked);
-  emit changed();
-}
-
-void HeatMapSetup::updateReverseScale(int option)
-{
-  source_.setReverseLogScale(option == Qt::Checked);
-  emit changed();
-}
-
-void HeatMapSetup::updateShowNumbers(int option)
-{
-  source_.setShowNumbers(option == Qt::Checked);
-  emit changed();
-}
-
-void HeatMapSetup::updateShowLegend(int option)
-{
-  source_.setShowLegend(option == Qt::Checked);
-  emit changed();
-}
-
-void HeatMapSetup::updateShowMinRange(int option)
-{
-  source_.setDrawBelowRangeMin(option == Qt::Checked);
-  emit changed();
-}
-
-void HeatMapSetup::updateShowMaxRange(int option)
-{
-  source_.setDrawAboveRangeMax(option == Qt::Checked);
-  emit changed();
-}
-
-void HeatMapSetup::updateRange()
-{
-  source_.setDisplayRange(source_.convertValueToPercent(min_range_selector_->value()),
-                          source_.convertValueToPercent(max_range_selector_->value()));
-  emit changed();
-}
-
-void HeatMapSetup::updateGridSize()
-{
-  if (!use_dbu_) {
-    source_.setGridSizes(grid_x_size_->value(), grid_y_size_->value());
-  } else {
-    const double dbu = dbu_;
-    source_.setGridSizes(grid_x_size_dbu_->value() / dbu, grid_x_size_dbu_->value() / dbu);
-  }
-  emit changed();
-}
-
-void HeatMapSetup::updateAlpha(int alpha)
-{
-  source_.setColorAlpha(alpha);
-  emit changed();
-}
-
-///////////
-
-HeatMapDataSource::HeatMapDataSource(const std::string& name,
+HeatMapDataSource::HeatMapDataSource(utl::Logger* logger,
+                                     const std::string& name,
                                      const std::string& short_name,
                                      const std::string& settings_group) :
     name_(name),
@@ -342,7 +57,7 @@ HeatMapDataSource::HeatMapDataSource(const std::string& name,
     colors_correct_(false),
     issue_redraw_(true),
     block_(nullptr),
-    logger_(nullptr),
+    logger_(logger),
     grid_x_size_(10.0),
     grid_y_size_(10.0),
     display_range_min_(getDisplayRangeMinimumValue()),
@@ -355,7 +70,7 @@ HeatMapDataSource::HeatMapDataSource(const std::string& name,
     show_numbers_(false),
     show_legend_(false),
     map_(),
-    renderer_(std::make_unique<HeatMapRenderer>(name_, *this)),
+    renderer_(std::make_unique<HeatMapRenderer>(*this)),
     setup_(nullptr),
     color_generator_(SpectrumGenerator(100.0))
 {
@@ -363,9 +78,14 @@ HeatMapDataSource::HeatMapDataSource(const std::string& name,
   updateMapColors();
 }
 
-void HeatMapDataSource::setLogger(utl::Logger* logger)
+HeatMapDataSource::~HeatMapDataSource()
 {
-  logger_ = logger;
+  Gui::get()->unregisterHeatMap(this);
+}
+
+void HeatMapDataSource::registerHeatMap()
+{
+  Gui::get()->registerHeatMap(this);
 }
 
 void HeatMapDataSource::redraw()
@@ -497,17 +217,47 @@ const std::string HeatMapDataSource::formatValue(double value, bool legend) cons
   return text.toStdString();
 }
 
+void HeatMapDataSource::addBooleanSetting(const std::string& name,
+                                          const std::string& label,
+                                          const std::function<bool(void)>& getter,
+                                          const std::function<void(bool)>& setter)
+{
+  settings_.push_back(MapSettingBoolean{name, label, getter, setter});
+}
+
+void HeatMapDataSource::addMultipleChoiceSetting(const std::string& name,
+                                                 const std::string& label,
+                                                 const std::function<std::vector<std::string>(void)>& choices,
+                                                 const std::function<std::string(void)>& getter,
+                                                 const std::function<void(std::string)>& setter)
+{
+  settings_.push_back(MapSettingMultiChoice{name, label, choices, getter, setter});
+}
+
 const Renderer::Settings HeatMapDataSource::getSettings() const
 {
-  return {{"DisplayMin", display_range_min_},
-          {"DisplayMax", display_range_max_},
-          {"GridX", grid_x_size_},
-          {"GridY", grid_y_size_},
-          {"Alpha", color_alpha_},
-          {"LogScale", log_scale_},
-          {"ReverseLog", reverse_log_},
-          {"ShowNumbers", show_numbers_},
-          {"ShowLegend", show_legend_}};
+  Renderer::Settings settings{
+    {"DisplayMin", display_range_min_},
+    {"DisplayMax", display_range_max_},
+    {"GridX", grid_x_size_},
+    {"GridY", grid_y_size_},
+    {"Alpha", color_alpha_},
+    {"LogScale", log_scale_},
+    {"ReverseLog", reverse_log_},
+    {"ShowNumbers", show_numbers_},
+    {"ShowLegend", show_legend_}};
+
+  for (const auto& setting : settings_) {
+    if (std::holds_alternative<MapSettingBoolean>(setting)) {
+      auto set = std::get<MapSettingBoolean>(setting);
+      settings[set.name] = set.getter();
+    } else if (std::holds_alternative<MapSettingMultiChoice>(setting)) {
+      auto set = std::get<MapSettingMultiChoice>(setting);
+      settings[set.name] = set.getter();
+    }
+  }
+
+  return settings;
 }
 
 void HeatMapDataSource::setSettings(const Renderer::Settings& settings)
@@ -521,6 +271,20 @@ void HeatMapDataSource::setSettings(const Renderer::Settings& settings)
   Renderer::setSetting<bool>(settings, "ReverseLog", reverse_log_);
   Renderer::setSetting<bool>(settings, "ShowNumbers", show_numbers_);
   Renderer::setSetting<bool>(settings, "ShowLegend", show_legend_);
+
+  for (const auto& setting : settings_) {
+    if (std::holds_alternative<MapSettingBoolean>(setting)) {
+      auto set = std::get<MapSettingBoolean>(setting);
+      bool temp_value = set.getter();
+      Renderer::setSetting<bool>(settings, set.name, temp_value);
+      set.setter(temp_value);
+    } else if (std::holds_alternative<MapSettingMultiChoice>(setting)) {
+      auto set = std::get<MapSettingMultiChoice>(setting);
+      std::string temp_value = set.getter();
+      Renderer::setSetting<std::string>(settings, set.name, temp_value);
+      set.setter(temp_value);
+    }
+  }
 
   // only reapply bounded value settings
   setDisplayRange(display_range_min_, display_range_max_);
@@ -719,12 +483,11 @@ void HeatMapDataSource::onHide()
 
 ///////////
 
-HeatMapRenderer::HeatMapRenderer(const std::string& display_control, HeatMapDataSource& datasource) :
-    display_control_(display_control),
+HeatMapRenderer::HeatMapRenderer(HeatMapDataSource& datasource) :
     datasource_(datasource),
     first_paint_(true)
 {
-  addDisplayControl(display_control_,
+  addDisplayControl(datasource_.getName(),
                     false,
                     [this]() { datasource_.showSetup(); },
                     {""}); // mutually exclusive to all
@@ -732,7 +495,7 @@ HeatMapRenderer::HeatMapRenderer(const std::string& display_control, HeatMapData
 
 void HeatMapRenderer::drawObjects(Painter& painter)
 {
-  if (!checkDisplayControl(display_control_)) {
+  if (!checkDisplayControl(datasource_.getName())) {
     if (!first_paint_) {
       first_paint_ = true; // reset check
       // first time so announce onHide
@@ -839,703 +602,14 @@ void HeatMapRenderer::setSettings(const Settings& settings)
   datasource_.setSettings(data_settings);
 }
 
-////////////
-
-RoutingCongestionDataSource::RoutingCongestionDataSource() :
-    HeatMapDataSource("Routing Congestion", "Routing", "RoutingCongestion"),
-    show_all_(true),
-    show_hor_(false),
-    show_ver_(false)
-{
-}
-
-void RoutingCongestionDataSource::makeAdditionalSetupOptions(QWidget* parent,
-                                                             QFormLayout* layout,
-                                                             const std::function<void(void)>& changed_callback)
-{
-  QComboBox* congestion_ = new QComboBox(parent);
-  congestion_->addItems({"All", "Horizontal", "Vertical"});
-
-  if (show_all_) {
-    congestion_->setCurrentIndex(0);
-  } else if (show_hor_) {
-    congestion_->setCurrentIndex(1);
-  } else if (show_ver_) {
-    congestion_->setCurrentIndex(2);
-  }
-
-  layout->addRow("Congestion Layers", congestion_);
-
-  QObject::connect(congestion_,
-                   QOverload<int>::of(&QComboBox::currentIndexChanged),
-                   [this](int value) {
-                     show_all_ = value == 0;
-                     show_hor_ = value == 1;
-                     show_ver_ = value == 2;
-                     destroyMap();
-                   });
-}
-
-double RoutingCongestionDataSource::getGridXSize() const
-{
-  if (getBlock() == nullptr) {
-    return default_grid_;
-  }
-
-  auto* gcellgrid = getBlock()->getGCellGrid();
-  if (gcellgrid == nullptr) {
-    return default_grid_;
-  }
-
-  std::vector<int> grid;
-  gcellgrid->getGridX(grid);
-
-  if (grid.size() < 2) {
-    return default_grid_;
-  } else {
-    const double delta = grid[1] - grid[0];
-    return delta / getBlock()->getDbUnitsPerMicron();
-  }
-}
-
-double RoutingCongestionDataSource::getGridYSize() const
-{
-  if (getBlock() == nullptr) {
-    return default_grid_;
-  }
-
-  auto* gcellgrid = getBlock()->getGCellGrid();
-  if (gcellgrid == nullptr) {
-    return default_grid_;
-  }
-
-  std::vector<int> grid;
-  gcellgrid->getGridY(grid);
-
-  if (grid.size() < 2) {
-    return default_grid_;
-  } else {
-    const double delta = grid[1] - grid[0];
-    return delta / getBlock()->getDbUnitsPerMicron();
-  }
-}
-
-bool RoutingCongestionDataSource::populateMap()
-{
-  if (getBlock() == nullptr) {
-    return false;
-  }
-
-  auto* grid = getBlock()->getGCellGrid();
-  if (grid == nullptr) {
-    return false;
-  }
-
-  auto gcell_congestion_data = grid->getCongestionMap();
-  if (gcell_congestion_data.empty()) {
-    return false;
-  }
-
-  std::vector<int> x_grid, y_grid;
-  grid->getGridX(x_grid);
-  const uint x_grid_sz = x_grid.size();
-  grid->getGridY(y_grid);
-  const uint y_grid_sz = y_grid.size();
-
-  for (const auto& [key, cong_data] : gcell_congestion_data) {
-    const uint x_idx = key.first;
-    const uint y_idx = key.second;
-
-    if (x_idx + 1 >= x_grid_sz || y_idx + 1 >= y_grid_sz) {
-      continue;
-    }
-
-    const odb::Rect gcell_rect(
-        x_grid[x_idx], y_grid[y_idx], x_grid[x_idx + 1], y_grid[y_idx + 1]);
-
-    const auto hor_capacity = cong_data.horizontal_capacity;
-    const auto hor_usage = cong_data.horizontal_usage;
-    const auto ver_capacity = cong_data.vertical_capacity;
-    const auto ver_usage = cong_data.vertical_usage;
-
-    //-1 indicates capacity is not well defined...
-    const double hor_congestion
-        = hor_capacity != 0 ? static_cast<double>(hor_usage) / hor_capacity : -1;
-    const double ver_congestion
-        = ver_capacity != 0 ? static_cast<double>(ver_usage) / ver_capacity : -1;
-
-    double congestion = 0.0;
-    if (show_all_) {
-      congestion = std::max(hor_congestion, ver_congestion);
-    } else if (show_hor_) {
-      congestion = hor_congestion;
-    } else {
-      congestion = ver_congestion;
-    }
-
-    if (congestion < 0) {
-      continue;
-    }
-
-    addToMap(gcell_rect, 100 * congestion);
-  }
-
-  return true;
-}
-
-void RoutingCongestionDataSource::combineMapData(bool base_has_value,
-                                                 double& base,
-                                                 const double new_data,
-                                                 const double data_area,
-                                                 const double intersection_area,
-                                                 const double rect_area)
-{
-  base += new_data * intersection_area / rect_area;
-}
-
-const Renderer::Settings RoutingCongestionDataSource::getSettings() const
-{
-  auto settings = HeatMapDataSource::getSettings();
-
-  settings["ShowAll"] = show_all_;
-  settings["ShowHor"] = show_hor_;
-  settings["ShowVer"] = show_ver_;
-
-  return settings;
-}
-
-void RoutingCongestionDataSource::setSettings(const Renderer::Settings& settings)
-{
-  HeatMapDataSource::setSettings(settings);
-
-  Renderer::setSetting<bool>(settings, "ShowAll", show_all_);
-  Renderer::setSetting<bool>(settings, "ShowHor", show_hor_);
-  Renderer::setSetting<bool>(settings, "ShowVer", show_ver_);
-}
-
-////////////
-
-PlacementDensityDataSource::PlacementDensityDataSource() :
-    HeatMapDataSource("Placement Density", "Placement", "PlacementDensity"),
-    include_taps_(true),
-    include_filler_(false),
-    include_io_(false)
-{
-}
-
-bool PlacementDensityDataSource::populateMap()
-{
-  if (getBlock() == nullptr) {
-    return false;
-  }
-
-  for (auto* inst : getBlock()->getInsts()) {
-    if (!inst->getPlacementStatus().isPlaced()) {
-      continue;
-    }
-    if (!include_filler_ && inst->getMaster()->isFiller()) {
-      continue;
-    }
-    if (!include_taps_ && (inst->getMaster()->getType() == odb::dbMasterType::CORE_WELLTAP || inst->getMaster()->isEndCap())) {
-      continue;
-    }
-    if (!include_io_ && (inst->getMaster()->isPad() || inst->getMaster()->isCover())) {
-      continue;
-    }
-    odb::Rect inst_box;
-    inst->getBBox()->getBox(inst_box);
-
-    addToMap(inst_box, 100.0);
-  }
-
-  return true;
-}
-
-void PlacementDensityDataSource::combineMapData(bool base_has_value,
-                                                double& base,
-                                                const double new_data,
-                                                const double data_area,
-                                                const double intersection_area,
-                                                const double rect_area)
-{
-  base += new_data * intersection_area / rect_area;
-}
-
-void PlacementDensityDataSource::makeAdditionalSetupOptions(QWidget* parent,
-                                                            QFormLayout* layout,
-                                                            const std::function<void(void)>& changed_callback)
-{
-  QCheckBox* taps = new QCheckBox(parent);
-  taps->setCheckState(include_taps_ ? Qt::Checked : Qt::Unchecked);
-  layout->addRow("Include taps and endcaps", taps);
-
-  QCheckBox* filler = new QCheckBox(parent);
-  filler->setCheckState(include_filler_ ? Qt::Checked : Qt::Unchecked);
-  layout->addRow("Include fillers", filler);
-
-  QCheckBox* io = new QCheckBox(parent);
-  io->setCheckState(include_io_ ? Qt::Checked : Qt::Unchecked);
-  layout->addRow("Include IO", io);
-
-  QObject::connect(taps,
-                   &QCheckBox::stateChanged,
-                   [this](int value) {
-                     include_taps_ = value == Qt::Checked;
-                     destroyMap();
-                   });
-
-  QObject::connect(filler,
-                   &QCheckBox::stateChanged,
-                   [this](int value) {
-                     include_filler_ = value == Qt::Checked;
-                     destroyMap();
-                   });
-
-  QObject::connect(io,
-                   &QCheckBox::stateChanged,
-                   [this](int value) {
-                     include_io_ = value == Qt::Checked;
-                     destroyMap();
-                   });
-}
-
-const Renderer::Settings PlacementDensityDataSource::getSettings() const
-{
-  auto settings = HeatMapDataSource::getSettings();
-
-  settings["Taps"] = include_taps_;
-  settings["Filler"] = include_filler_;
-  settings["IO"] = include_io_;
-
-  return settings;
-}
-
-void PlacementDensityDataSource::setSettings(const Renderer::Settings& settings)
-{
-  HeatMapDataSource::setSettings(settings);
-
-  Renderer::setSetting<bool>(settings, "Taps", include_taps_);
-  Renderer::setSetting<bool>(settings, "Filler", include_filler_);
-  Renderer::setSetting<bool>(settings, "IO", include_io_);
-}
-
-void PlacementDensityDataSource::onShow()
-{
-  HeatMapDataSource::onShow();
-
-  addOwner(getBlock());
-}
-
-void PlacementDensityDataSource::onHide()
-{
-  HeatMapDataSource::onHide();
-
-  removeOwner();
-}
-
-void PlacementDensityDataSource::inDbInstCreate(odb::dbInst*)
-{
-  destroyMap();
-}
-
-void PlacementDensityDataSource::inDbInstCreate(odb::dbInst*, odb::dbRegion*)
-{
-  destroyMap();
-}
-
-void PlacementDensityDataSource::inDbInstDestroy(odb::dbInst*)
-{
-  destroyMap();
-}
-
-void PlacementDensityDataSource::inDbInstPlacementStatusBefore(odb::dbInst*, const odb::dbPlacementStatus&)
-{
-  destroyMap();
-}
-
-void PlacementDensityDataSource::inDbInstSwapMasterBefore(odb::dbInst*, odb::dbMaster*)
-{
-  destroyMap();
-}
-
-void PlacementDensityDataSource::inDbInstSwapMasterAfter(odb::dbInst*)
-{
-  destroyMap();
-}
-
-void PlacementDensityDataSource::inDbPreMoveInst(odb::dbInst*)
-{
-  destroyMap();
-}
-
-void PlacementDensityDataSource::inDbPostMoveInst(odb::dbInst*)
-{
-  destroyMap();
-}
-
-////////////
-
-PowerDensityDataSource::PowerDensityDataSource() :
-    RealValueHeatMapDataSource("W", "Power Density", "Power", "PowerDensity"),
-    sta_(nullptr),
-    include_internal_(true),
-    include_leakage_(true),
-    include_switching_(true),
-    corner_(nullptr)
-{
-  setIssueRedraw(false); // disable during initial setup
-  setLogScale(true);
-  setIssueRedraw(true);
-}
-
-bool PowerDensityDataSource::populateMap()
-{
-  if (getBlock() == nullptr || sta_ == nullptr) {
-    return false;
-  }
-
-  if (sta_->cmdNetwork() == nullptr) {
-    return false;
-  }
-
-  ensureCorner();
-
-  auto* network = sta_->getDbNetwork();
-
-  const bool include_all = include_internal_ && include_leakage_ && include_switching_;
-  for (auto* inst : getBlock()->getInsts()) {
-    if (!inst->getPlacementStatus().isPlaced()) {
-      continue;
-    }
-
-    sta::PowerResult power;
-    sta_->power(network->dbToSta(inst), corner_, power);
-
-    float pwr = 0.0;
-    if (include_all) {
-      pwr = power.total();
-    } else {
-      if (include_internal_) {
-        pwr += power.internal();
-      }
-      if (include_leakage_) {
-        pwr += power.switching();
-      }
-      if (include_switching_) {
-        pwr += power.leakage();
-      }
-    }
-
-    odb::Rect inst_box;
-    inst->getBBox()->getBox(inst_box);
-
-    addToMap(inst_box, pwr);
-  }
-
-  return true;
-}
-
-void PowerDensityDataSource::combineMapData(bool base_has_value,
-                                            double& base,
-                                            const double new_data,
-                                            const double data_area,
-                                            const double intersection_area,
-                                            const double rect_area)
-{
-  base += (new_data / data_area) * intersection_area;
-}
-
-void PowerDensityDataSource::makeAdditionalSetupOptions(QWidget* parent,
-                                                        QFormLayout* layout,
-                                                        const std::function<void(void)>& changed_callback)
-{
-  ensureCorner();
-
-  QComboBox* corners = nullptr;
-  if (sta_->multiCorner()) {
-    corners = new QComboBox(parent);
-    int selection = 0;
-    for (auto* corner : *sta_->corners()) {
-      corners->addItem(corner->name());
-      if (corner_ == corner) {
-        selection = corners->count() - 1;
-      }
-    }
-    corners->setCurrentIndex(selection);
-    layout->addRow("Corner", corners);
-  }
-
-  QCheckBox* internal = new QCheckBox(parent);
-  internal->setCheckState(include_internal_ ? Qt::Checked : Qt::Unchecked);
-  layout->addRow("Internal power", internal);
-
-  QCheckBox* switching = new QCheckBox(parent);
-  switching->setCheckState(include_switching_ ? Qt::Checked : Qt::Unchecked);
-  layout->addRow("Switching power", switching);
-
-  QCheckBox* leakage = new QCheckBox(parent);
-  leakage->setCheckState(include_leakage_ ? Qt::Checked : Qt::Unchecked);
-  layout->addRow("Leakage power", leakage);
-
-  QObject::connect(internal,
-                   &QCheckBox::stateChanged,
-                   [this, changed_callback](int value) {
-                     include_internal_ = value == Qt::Checked;
-                     destroyMap();
-                     changed_callback();
-                   });
-
-  QObject::connect(switching,
-                   &QCheckBox::stateChanged,
-                   [this, changed_callback](int value) {
-                     include_switching_ = value == Qt::Checked;
-                     destroyMap();
-                     changed_callback();
-                   });
-
-  QObject::connect(leakage,
-                   &QCheckBox::stateChanged,
-                   [this, changed_callback](int value) {
-                     include_leakage_ = value == Qt::Checked;
-                     destroyMap();
-                     changed_callback();
-                   });
-
-  if (corners != nullptr) {
-    QObject::connect(corners,
-                     &QComboBox::currentTextChanged,
-                     [this, changed_callback](const QString& corner) {
-                       setCorner(corner.toStdString());
-                       destroyMap();
-                       changed_callback();
-                     });
-  }
-}
-
-const Renderer::Settings PowerDensityDataSource::getSettings() const
-{
-  auto settings = HeatMapDataSource::getSettings();
-
-  settings["Internal"] = include_internal_;
-  settings["Switching"] = include_switching_;
-  settings["Leakage"] = include_leakage_;
-  if (corner_ != nullptr) {
-    settings["corner"] = std::string(corner_->name());
-  }
-
-  return settings;
-}
-
-void PowerDensityDataSource::setSettings(const Renderer::Settings& settings)
-{
-  ensureCorner();
-
-  HeatMapDataSource::setSettings(settings);
-
-  Renderer::setSetting<bool>(settings, "Internal", include_internal_);
-  Renderer::setSetting<bool>(settings, "Switching", include_switching_);
-  Renderer::setSetting<bool>(settings, "Leakage", include_leakage_);
-
-  std::string corner = corner_->name();
-  Renderer::setSetting<std::string>(settings, "corner", corner);
-  setCorner(corner);
-}
-
-void PowerDensityDataSource::ensureCorner()
-{
-  if (corner_ != nullptr) {
-    return;
-  }
-
-  auto corners = sta_->corners()->corners();
-  corner_ = corners[0];
-}
-
-void PowerDensityDataSource::setCorner(const std::string& name)
-{
-  corner_ = sta_->findCorner(name.c_str());
-}
-
-////////////
-
-IRDropDataSource::IRDropDataSource() :
-    RealValueHeatMapDataSource("V", "IR Drop", "IRDrop", "IRDrop"),
-    psm_(nullptr),
-    tech_(nullptr),
-    layer_(nullptr)
-{
-}
-
-void IRDropDataSource::setBlock(odb::dbBlock* block)
-{
-  HeatMapDataSource::setBlock(block);
-  if (block != nullptr) {
-    tech_ = block->getDb()->getTech();
-  }
-}
-
-double IRDropDataSource::getGridSizeMinimumValue() const
-{
-  odb::dbBlock* block = getBlock();
-  if (block == nullptr || psm_ == nullptr) {
-    return RealValueHeatMapDataSource::getGridSizeMinimumValue();
-  }
-
-  try {
-    const double resolution = psm_->getMinimumResolution();
-    double resolution_um = resolution / block->getDbUnitsPerMicron();
-    if (resolution_um > getGridSizeMaximumValue()) {
-      resolution_um = RealValueHeatMapDataSource::getGridSizeMinimumValue();
-    }
-    return resolution_um;
-  } catch (const std::runtime_error& /* e */) {
-    // psm is not setup up
-    return RealValueHeatMapDataSource::getGridSizeMinimumValue();
-  }
-}
-
-bool IRDropDataSource::populateMap()
-{
-  if (getBlock() == nullptr || psm_ == nullptr || tech_ == nullptr) {
-    return false;
-  }
-
-  ensureLayer();
-
-  std::map<odb::dbTechLayer*, std::map<odb::Point, double>> ir_drops;
-  psm_->getIRDropMap(ir_drops);
-
-  if (ir_drops.empty()) {
-    return false;
-  }
-
-  // track min/max here to make it constant across all layers
-  double min = std::numeric_limits<double>::max();
-  double max = std::numeric_limits<double>::min();
-  for (const auto& [layer, drop_map] : ir_drops) {
-    for (const auto& [point, drop] : drop_map) {
-      min = std::min(min, drop);
-      max = std::max(max, drop);
-    }
-  }
-  setMinValue(min);
-  setMaxValue(max);
-
-  auto& ir_drop = ir_drops[layer_];
-  for (const auto& [point, drop] : ir_drop) {
-    addToMap({point, point}, drop);
-  }
-
-  return true;
-}
-
-void IRDropDataSource::determineMinMax(const HeatMapDataSource::Map& map)
-{
-  // do nothing handled in populateMap
-}
-
-void IRDropDataSource::combineMapData(bool base_has_value,
-                                      double& base,
-                                      const double new_data,
-                                      const double data_area,
-                                      const double intersection_area,
-                                      const double rect_area)
-{
-  if (!base_has_value) {
-    base = new_data;
-  } else {
-    base = std::max(base, new_data);
-  }
-}
-
-void IRDropDataSource::makeAdditionalSetupOptions(QWidget* parent,
-                                                  QFormLayout* layout,
-                                                  const std::function<void(void)>& changed_callback)
-{
-  if (tech_ == nullptr) {
-    return;
-  }
-
-  ensureLayer();
-
-  QComboBox* layers = new QComboBox(parent);
-  int layer_selection = 0;
-  for (auto* layer : tech_->getLayers()) {
-    if (layer->getRoutingLevel() == 0) {
-      continue;
-    }
-
-    layers->addItem(layer->getConstName());
-    if (layer_ == layer) {
-      layer_selection = layers->count() - 1;
-    }
-  }
-  layers->setCurrentIndex(layer_selection);
-  layout->addRow("Layer", layers);
-
-  QObject::connect(layers,
-                   &QComboBox::currentTextChanged,
-                   [this, changed_callback](const QString& layer) {
-                     setLayer(layer.toStdString());
-                     destroyMap();
-                     changed_callback();
-                   });
-}
-
-const Renderer::Settings IRDropDataSource::getSettings() const
-{
-  auto settings = HeatMapDataSource::getSettings();
-
-  if (layer_ != nullptr) {
-    settings["layer"] = std::string(layer_->getName());
-  }
-
-  return settings;
-}
-
-void IRDropDataSource::setSettings(const Renderer::Settings& settings)
-{
-  ensureLayer();
-
-  HeatMapDataSource::setSettings(settings);
-
-  std::string layer = "";
-  if (layer_ != nullptr) {
-    layer = layer_->getName();
-  }
-  Renderer::setSetting<std::string>(settings, "layer", layer);
-  setLayer(layer);
-}
-
-void IRDropDataSource::ensureLayer()
-{
-  if (layer_ != nullptr) {
-    return;
-  }
-
-  if (tech_ == nullptr) {
-    return;
-  }
-
-  layer_ = tech_->findRoutingLayer(1);
-}
-
-void IRDropDataSource::setLayer(const std::string& name)
-{
-  if (tech_ == nullptr) {
-    return;
-  }
-
-  layer_ = tech_->findLayer(name.c_str());
-}
-
 //////////
 
-RealValueHeatMapDataSource::RealValueHeatMapDataSource(const std::string& unit_suffix,
+RealValueHeatMapDataSource::RealValueHeatMapDataSource(utl::Logger* logger,
+                                                       const std::string& unit_suffix,
                                                        const std::string& name,
                                                        const std::string& short_name,
                                                        const std::string& settings_group) :
-    HeatMapDataSource(name, short_name, settings_group),
+    HeatMapDataSource(logger, name, short_name, settings_group),
     unit_suffix_(unit_suffix),
     units_(unit_suffix_),
     min_(0.0),
