+++ src/gui/src/mainWindow.cpp
@@ -61,6 +61,7 @@
 #include "utl/Logger.h"
 #include "timingWidget.h"
 #include "drcWidget.h"
+#include "gui/heatMap.h"
 
 // must be loaded in global namespace
 static void loadQTResources()
@@ -325,8 +326,9 @@ MainWindow::MainWindow(QWidget* parent)
 
 MainWindow::~MainWindow()
 {
+  auto* gui = Gui::get();
   // unregister descriptors
-  Gui::get()->unregisterDescriptor<Ruler*>();
+  gui->unregisterDescriptor<Ruler*>();
 }
 
 void MainWindow::setDatabase(odb::dbDatabase* db)
@@ -343,12 +345,12 @@ void MainWindow::setDatabase(odb::dbDatabase* db)
 
 void MainWindow::setBlock(odb::dbBlock* block)
 {
-  for (auto* heat_map : getHeatMaps()) {
+  for (auto* heat_map : Gui::get()->getHeatMaps()) {
     heat_map->setBlock(block);
   }
 }
 
-void MainWindow::init(sta::dbSta* sta, psm::PDNSim* psm)
+void MainWindow::init(sta::dbSta* sta)
 {
   // Setup timing widget
   timing_widget_->init(sta);
@@ -365,13 +367,6 @@ void MainWindow::init(sta::dbSta* sta, psm::PDNSim* psm)
   gui->registerDescriptor<odb::dbTechLayer*>(new DbTechLayerDescriptor(db_));
   gui->registerDescriptor<DbItermAccessPoint>(new DbItermAccessPointDescriptor(db_));
   gui->registerDescriptor<Ruler*>(new RulerDescriptor(rulers_, db_));
-
-  // renderers
-  power_density_data_.setSTA(sta);
-  ir_drop_data_.setPSM(psm);
-  for (auto* heat_map : getHeatMaps()) {
-    gui->registerRenderer(heat_map->getRenderer());
-  }
 }
 
 void MainWindow::createStatusBar()
@@ -451,7 +446,7 @@ void MainWindow::createActions()
 
 void MainWindow::setUseDBU(bool use_dbu)
 {
-  for (auto* heat_map : getHeatMaps()) {
+  for (auto* heat_map : Gui::get()->getHeatMaps()) {
     heat_map->setUseDBU(use_dbu);
   }
 }
@@ -471,10 +466,9 @@ void MainWindow::createMenus()
   tools_menu_ = menuBar()->addMenu("&Tools");
   tools_menu_->addAction(build_ruler_);
   auto heat_maps = tools_menu_->addMenu("&Heat maps");
-  for (auto* heat_map : getHeatMaps()) {
-    connect(heat_maps->addAction(QString::fromStdString(heat_map->getName())),
-            &QAction::triggered,
-            [heat_map]() { heat_map->showSetup(); });
+  heat_maps->setObjectName("HeatMaps");
+  for (auto* heat_map : Gui::get()->getHeatMaps()) {
+    registerHeatMap(heat_map);
   }
 
   windows_menu_ = menuBar()->addMenu("&Windows");
@@ -1071,11 +1065,6 @@ void MainWindow::setLogger(utl::Logger* logger)
   script_->setLogger(logger);
   viewer_->setLogger(logger);
   drc_viewer_->setLogger(logger);
-
-  // heat maps
-  for (auto* heat_map : getHeatMaps()) {
-    heat_map->setLogger(logger);
-  }
 }
 
 void MainWindow::fit()
@@ -1162,94 +1151,6 @@ const std::vector<std::string> MainWindow::getRestoreTclCommands()
   return cmds;
 }
 
-const std::vector<HeatMapDataSource*> MainWindow::getHeatMaps()
-{
-  return {
-    &routing_congestion_data_,
-    &placement_density_data_,
-    &power_density_data_,
-    &ir_drop_data_
-  };
-}
-
-void MainWindow::setHeatMapSetting(const std::string& name, const std::string& option, const Renderer::Setting& value)
-{
-  HeatMapDataSource* source = nullptr;
-
-  for (auto* heat_map : getHeatMaps()) {
-    if (heat_map->getShortName() == name) {
-      source = heat_map;
-      break;
-    }
-  }
-
-  if (source == nullptr) {
-    QStringList options;
-    for (const auto* heat_map : getHeatMaps()) {
-      options.append(QString::fromStdString(heat_map->getShortName()));
-    }
-    logger_->error(utl::GUI, 28, "{} is not a known map. Valid options are: {}", name, options.join(", ").toStdString());
-  }
-
-  const std::string rebuild_map_option = "rebuild";
-  if (option == rebuild_map_option) {
-    source->destroyMap();
-  } else {
-    auto settings = source->getSettings();
-
-    if (settings.count(option) == 0) {
-      QStringList options;
-      options.append(QString::fromStdString(rebuild_map_option));
-      for (const auto& [key, kv] : settings) {
-        options.append(QString::fromStdString(key));
-      }
-      logger_->error(utl::GUI, 29, "{} is not a valid option. Valid options are: {}", option, options.join(", ").toStdString());
-    }
-
-    auto current_value = settings[option];
-    if (std::holds_alternative<bool>(current_value)) {
-      // is bool
-      if (auto* s = std::get_if<bool>(&value)) {
-        settings[option] = *s;
-      } if (auto* s = std::get_if<int>(&value)) {
-        settings[option] = *s != 0;
-      } if (auto* s = std::get_if<double>(&value)) {
-        settings[option] = *s != 0.0;
-      } else {
-        logger_->error(utl::GUI, 60, "{} must be a boolean", option);
-      }
-    } else if (std::holds_alternative<int>(current_value)) {
-      // is int
-      if (auto* s = std::get_if<int>(&value)) {
-        settings[option] = *s;
-      } else if (auto* s = std::get_if<double>(&value)) {
-        settings[option] = static_cast<int>(*s);
-      } else {
-        logger_->error(utl::GUI, 61, "{} must be an integer or double", option);
-      }
-    } else if (std::holds_alternative<double>(current_value)) {
-      // is double
-      if (auto* s = std::get_if<int>(&value)) {
-        settings[option] = static_cast<double>(*s);
-      } else if (auto* s = std::get_if<double>(&value)) {
-        settings[option] = *s;
-      } else {
-        logger_->error(utl::GUI, 62, "{} must be an integer or double", option);
-      }
-    } else {
-      // is string
-      if (auto* s = std::get_if<std::string>(&value)) {
-        settings[option] = *s;
-      } else {
-        logger_->error(utl::GUI, 63, "{} must be a string", option);
-      }
-    }
-    source->setSettings(settings);
-  }
-
-  source->getRenderer()->redraw();
-}
-
 std::string MainWindow::convertDBUToString(int value, bool add_units) const
 {
   if (show_dbu_->isChecked()) {
@@ -1311,6 +1212,23 @@ void MainWindow::timingCone(std::variant<odb::dbITerm*, odb::dbBTerm*> term, boo
   } else {
     renderer->setBTerm(std::get<odb::dbBTerm*>(term), fanin, fanout);
   }
+}
+
+void MainWindow::registerHeatMap(HeatMapDataSource* heatmap)
+{
+  auto* heat_maps = menuBar()->findChild<QMenu*>("HeatMaps");
+  auto* action = heat_maps->addAction(QString::fromStdString(heatmap->getName()));
+  heatmap_actions_[heatmap] = action;
+  connect(action,
+          &QAction::triggered,
+          [heatmap]() { heatmap->showSetup(); });
+}
+
+void MainWindow::unregisterHeatMap(HeatMapDataSource* heatmap)
+{
+  auto* heat_maps = menuBar()->findChild<QMenu*>("HeatMaps");
+  heat_maps->removeAction(heatmap_actions_[heatmap]);
+  heatmap_actions_.erase(heatmap);
 }
 
 }  // namespace gui
